Extract sync UI helpers
